requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
gunicorn==21.2.0
load-dotenv==0.1.0
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.1
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2023.3
